requirements.txt
Django==2.0.5
gunicorn
psycopg2==2.8.4
django-heroku
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0